+++ app/src/main/java/ru/gb/car/entity/Point.kt
@@ -3,4 +3,5 @@ package ru.gb.car.entity
 class Point {
     var x: Float = 100f
     var y: Float = 100f
+    var angle: Float = 0.0f
 }
